requirements.txt
dbt-core==1.7.0
dbt-snowflake==1.7.0
pandas==2.2.3
snowflake-connector-python==3.14.0